add miner address
