+++ src/components/TransactionPage/components/SubmissionBoxes.tsx
@@ -1,0 +1,29 @@
+const SubmissionBoxes = {
+  flashbots: (
+    <span className="m-1 p-4 rounded-3xl py-2.5 bg-purple inline-block text-xs text-bold text-blue-light shadow-sm font-bold text-center">
+      Flashbots
+    </span>
+  ),
+  eden: (
+    <span className="m-1 p-4 rounded-3xl py-2.5 bg-green inline-block text-xs text-bold text-blue-light shadow-sm font-bold text-center">
+      Eden
+    </span>
+  ),
+  agg: (
+    <span className="m-1 p-4 rounded-3xl py-2.5 bg-green inline-block text-xs text-bold text-blue-light shadow-sm font-bold text-center">
+      Eden Aggregator
+    </span>
+  ),
+  ethermine: (
+    <span className="m-1 p-4 rounded-3xl py-2.5 bg-orange inline-block text-xs text-bold text-blue-light shadow-sm font-bold text-center">
+      Ethermine
+    </span>
+  ),
+  unknown: (
+    <span className="p-4 rounded-3xl py-2.5 bg-white inline-block text-xs text-bold text-blue-light shadow-sm font-bold text-center">
+      Unknown
+    </span>
+  ),
+};
+
+export default SubmissionBoxes;
