+++ src/components/TransactionPage/index.tsx
@@ -1,0 +1,3 @@
+import TransactionPage from './TransactionPage';
+
+export default TransactionPage;
